hw77

+++ Binomial.java
@@ -25,11 +25,11 @@ public class Binomial {
     }
 
     // Computes the Binomial function, efficiently
-    public static int binomial(int n, int k) {
+    public static long binomial(int n, int k) {
         //// This function creates a 2D array, say memo, 
 		//// and then initializes all its elements to -1.
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
-		int[][] newMemo = new int[n + 1][k + 1];
+		long[][] newMemo = new long[n + 1][k + 1];
         for (int i = 0; i <= n; i++) {
             for (int j = 0; j <= k; j++) {
                 newMemo[i][j] = -1;
@@ -40,7 +40,7 @@ public class Binomial {
         return binomial(n, k, newMemo);
     }
 
-    private static int binomial(int n, int k, int[][] memo) {
+    private static long binomial(int n, int k, long[][] memo) {
         if (memo[n][k] != -1) {
             return memo[n][k];
         }
